rename, add doc comments, and remove from openai tests
