+++ js/packages/openinference-core/src/trace/index.ts
@@ -1,3 +1,3 @@
 export * from "./contextAttributes";
 export * from "./types";
-export * from "./trace_config";
+export * from "./trace-config";
