+++ js/packages/openinference-core/src/trace/trace-config/OISpan.ts
