+++ js/packages/openinference-core/src/trace/trace-config/OITracer.ts
@@ -45,9 +45,17 @@ function formatStartActiveSpanParams<F extends OpenInferenceActiveSpanCallback>(
   return { opts, ctx, fn };
 }
 
+/**
+ * A wrapper around the OpenTelemetry {@link Tracer} interface that masks sensitive information based on the passed in {@link TraceConfig}.
+ */
 export class OITracer implements Tracer {
   private readonly tracer: Tracer;
   private readonly config: TraceConfig;
+  /**
+   *
+   * @param tracer The OpenTelemetry {@link Tracer} to wrap
+   * @param traceConfig The {@link TraceConfigOptions} to set to control the behavior of the tracer
+   */
   constructor({
     tracer,
     traceConfig,
